requirements.txt
faiss-cpu==1.7.4
numpy==1.26.4
google-generativeai==0.8.5
grpcio-status==1.71.2
importlib-metadata==8.0.0
librosa==0.10.1
onnxruntime==1.23.2
openai-whisper==20231117
opencv-python==4.9.0.80
pygame==2.6.1
python-dotenv==1.2.1
rapidfuzz==3.5.2
rembg==2.0.69
sounddevice==0.4.6
tomli==2.0.1
torchaudio>=2.0.0
torchvision>=0.15.0
transformers==4.57.3
